More refactoring. Added frequency step to calculix.
